requirements.txt
numpy
matplotlib
scipy
cupy-cuda12x
torch
